+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""BugCapsule test package."""